+++ html/gfx.html
@@ -1,79 +1,30 @@
 <h1>Graphics Design</h1>
-<a
-  class="card hover"
-  href="https://discord.gg/qsociety"
-  alt="Quantum Society Discord server"
-  target="_blank"
-  rel="noopener noreferrer"
-  style="background-color: rgba(29, 91, 138, 0.25)"
->
-  <img
-    alt="Quantum Society Icon"
-    class="card-icon circle"
-    src="assets/projects/quantum-society.webp"
-  />
+<a class="card hover" href="https://discord.gg/qsociety" alt="Quantum Society Discord server" target="_blank"
+  rel="noopener noreferrer" style="background-color: rgba(29, 91, 138, 0.25)">
+  <img alt="Quantum Society Icon" class="card-icon circle" src="assets/projects/quantum-society.webp" />
   <div class="card-content">
     <h3>Quantum Society</h3>
     <p>
-      My Discord server, all art in it including the penguin emotes (<img
-        alt="Penguin Emoji"
-        class="emoji"
-        src="assets/emojis/penguin.webp"
-      />
-      <img
-        alt="Penguin Heart Emoji"
-        class="emoji"
-        src="assets/emojis/penguin-heart.webp"
-      />
-      <img
-        alt="Penguin Spin Emoji"
-        class="emoji"
-        src="assets/emojis/penguin-spin.webp"
-      />
-      <img
-        alt="Penguin Party Emoji"
-        class="emoji"
-        src="assets/emojis/penguin-party.webp"
-      />
-      <img
-        alt="Penguin Vibe Emoji"
-        class="emoji"
-        src="assets/emojis/penguin-vibe.webp"
-      />)
+      My Discord server, all art in it including the penguin emotes (<img alt="Penguin Emoji" class="emoji"
+        src="assets/emojis/penguin.webp?width=25&height=25" />
+      <img alt="Penguin Heart Emoji" class="emoji" src="assets/emojis/penguin-heart.webp?width=25&height=25" />
+      <img alt="Penguin Spin Emoji" class="emoji" src="assets/emojis/penguin-spin.webp?width=25&height=25" />
+      <img alt="Penguin Party Emoji" class="emoji" src="assets/emojis/penguin-party.webp?width=25&height=25" />
+      <img alt="Penguin Vibe Emoji" class="emoji" src="assets/emojis/penguin-vibe.webp?width=25&height=25" />)
     </p>
   </div>
 </a>
-<a
-  class="card hover"
-  href="https://discord.gg/ps"
-  alt="PlayStation Discord server"
-  target="_blank"
-  rel="noopener noreferrer"
-  style="background-color: rgba(8, 15, 138, 0.25)"
->
-  <img
-    alt="PlayStation Discord Icon"
-    class="card-icon circle"
-    src="assets/projects/playstation.webp"
-  />
+<a class="card hover" href="https://discord.gg/ps" alt="PlayStation Discord server" target="_blank"
+  rel="noopener noreferrer" style="background-color: rgba(8, 15, 138, 0.25)">
+  <img alt="PlayStation Discord Icon" class="card-icon circle" src="assets/projects/playstation.webp" />
   <div class="card-content">
     <h3>PlayStation Discord</h3>
     <p>The icon for the community run PlayStation Discord server</p>
   </div>
 </a>
-<a
-  class="card hover"
-  href="https://discord.gg/ian"
-  alt="Sol Sanctum Discord server"
-  target="_blank"
-  rel="noopener noreferrer"
-  style="background-color: rgba(171, 55, 206, 0.25)"
->
-  <img
-    alt="Sol Sanctum Discord Icon"
-    class="card-icon circle"
-    src="assets/projects/ian.webp"
-  />
+<a class="card hover" href="https://discord.gg/ian" alt="Sol Sanctum Discord server" target="_blank"
+  rel="noopener noreferrer" style="background-color: rgba(171, 55, 206, 0.25)">
+  <img alt="Sol Sanctum Discord Icon" class="card-icon circle" src="assets/projects/ian.webp" />
   <div class="card-content">
     <h3>Sol Sanctum Discord</h3>
     <p>The icon and some stickers for the Sol Sanctum Discord server</p>
