+++ html/index.html
@@ -3,10 +3,10 @@
 
 <head>
   <style>
-    @import "../css/index.css";
+    @import "/css/index.css";
   </style>
   <script type="module">
-    import '../js/index.ts'
+    import '/js/index.ts'
   </script>
 
   <title>Commandtechno</title>
