+++ scripts/postbuild.ts
@@ -1,0 +1,3 @@
+import { existsSync, rmSync } from "fs";
+
+if (existsSync("temp.html")) rmSync("temp.html");
